Move data and animations
