+++ frontend/src/Projects.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import projects from "./projects.json";
+import projects from "./data/projects.json";
 import "./Projects.css";
 import portfolioImg from './assets/projects/portfolio.png';
 import comingSoonImg from './assets/projects/comingsoon.jpeg';
