+++ frontend/src/assets/animations/map_pointer.json
