+++ frontend/src/data/projects.json
